pyproject.toml
[project]
name = "lucinka"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-sqlalchemy>=3.1.1",
    "marshmallow>=4.0.1",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D"]

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.hatch.version]
path = "lucinka/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["lucinka"]

[build-system]
requires = ["hatchling==1.27.0"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest==8.4.1",
    "pytest-snapshot==0.9.0",
    "pytest-cov==6.2.1",
    "ruff==0.12.4",
    "ty==0.0.1a16",
]
